@@ -1,10 +1,10 @@
 name: flutter_reactive_ble_example
 description: Demonstrates how to use the flutter_reactive_ble plugin.
-version: 2.7.1
+version: 1.0.0-nullsafety.0
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0-133.2.beta <3.0.0'
   flutter: ">=1.10.0"
 
 dependencies:
@@ -12,10 +12,10 @@ dependencies:
     sdk: flutter
   flutter_reactive_ble:
     path: ../
-  provider: ^4.0.4
+  provider: ^5.0.0-nullsafety.2
 
 dev_dependencies:
-  dependency_validator:
+  dependency_validator: ^2.0.1
   flutter_test:
     sdk: flutter
 
