@@ -8,9 +8,10 @@ import 'device_detail_screen.dart';
 
 class DeviceListScreen extends StatelessWidget {
   @override
-  Widget build(BuildContext context) => Consumer2<BleScanner, BleScannerState>(
+  Widget build(BuildContext context) => Consumer2<BleScanner, BleScannerState?>(
         builder: (_, bleScanner, bleScannerState, __) => _DeviceList(
-          scannerState: bleScannerState,
+          scannerState: bleScannerState ??
+              BleScannerState(discoveredDevices: [], scanIsInProgress: false),
           startScan: bleScanner.startScan,
           stopScan: bleScanner.stopScan,
         ),
@@ -19,12 +20,9 @@ class DeviceListScreen extends StatelessWidget {
 
 class _DeviceList extends StatefulWidget {
   const _DeviceList(
-      {@required this.scannerState,
-      @required this.startScan,
-      @required this.stopScan})
-      : assert(scannerState != null),
-        assert(startScan != null),
-        assert(stopScan != null);
+      {required this.scannerState,
+      required this.startScan,
+      required this.stopScan});
 
   final BleScannerState scannerState;
   final void Function(List<Uuid>) startScan;
@@ -35,7 +33,7 @@ class _DeviceList extends StatefulWidget {
 }
 
 class _DeviceListState extends State<_DeviceList> {
-  TextEditingController _uuidController;
+  TextEditingController? _uuidController;
 
   @override
   void initState() {
@@ -47,12 +45,12 @@ class _DeviceListState extends State<_DeviceList> {
   @override
   void dispose() {
     widget.stopScan();
-    _uuidController.dispose();
+    _uuidController!.dispose();
     super.dispose();
   }
 
   bool _isValidUuidInput() {
-    final uuidText = _uuidController.text;
+    final uuidText = _uuidController!.text;
     if (uuidText.isEmpty) {
       return true;
     } else {
@@ -66,8 +64,8 @@ class _DeviceListState extends State<_DeviceList> {
   }
 
   void _startScanning() {
-    final text = _uuidController.text;
-    widget.startScan(text.isEmpty ? [] : [Uuid.parse(_uuidController.text)]);
+    final text = _uuidController!.text;
+    widget.startScan(text.isEmpty ? [] : [Uuid.parse(_uuidController!.text)]);
   }
 
   @override
@@ -89,7 +87,7 @@ class _DeviceListState extends State<_DeviceList> {
                     enabled: !widget.scannerState.scanIsInProgress,
                     decoration: InputDecoration(
                         errorText:
-                            _uuidController.text.isEmpty || _isValidUuidInput()
+                            _uuidController!.text.isEmpty || _isValidUuidInput()
                                 ? null
                                 : 'Invalid UUID format'),
                     autocorrect: false,
