@@ -1,13 +1,12 @@
 import 'package:flutter_reactive_ble/flutter_reactive_ble.dart';
 import 'package:flutter_reactive_ble/src/generated/bledata.pb.dart' as pb;
 import 'package:flutter_reactive_ble/src/model/uuid.dart';
-import 'package:meta/meta.dart';
 
 abstract class ArgsToProtobufConverter {
   pb.ConnectToDeviceRequest createConnectToDeviceArgs(
     String id,
-    Map<Uuid, List<Uuid>> servicesWithCharacteristicsToDiscover,
-    Duration connectionTimeout,
+    Map<Uuid, List<Uuid>>? servicesWithCharacteristicsToDiscover,
+    Duration? connectionTimeout,
   );
 
   pb.DisconnectFromDeviceRequest createDisconnectDeviceArgs(String deviceId);
@@ -40,9 +39,9 @@ abstract class ArgsToProtobufConverter {
   );
 
   pb.ScanForDevicesRequest createScanForDevicesRequest({
-    @required List<Uuid> withServices,
-    @required ScanMode scanMode,
-    @required bool requireLocationServicesEnabled,
+    required List<Uuid> withServices,
+    required ScanMode? scanMode,
+    required bool? requireLocationServicesEnabled,
   });
 
   pb.ClearGattCacheRequest createClearGattCacheRequest(String deviceId);
@@ -56,8 +55,8 @@ class ArgsToProtobufConverterImpl implements ArgsToProtobufConverter {
   @override
   pb.ConnectToDeviceRequest createConnectToDeviceArgs(
     String id,
-    Map<Uuid, List<Uuid>> servicesWithCharacteristicsToDiscover,
-    Duration connectionTimeout,
+    Map<Uuid, List<Uuid>>? servicesWithCharacteristicsToDiscover,
+    Duration? connectionTimeout,
   ) {
     final args = pb.ConnectToDeviceRequest()..deviceId = id;
 
@@ -69,7 +68,7 @@ class ArgsToProtobufConverterImpl implements ArgsToProtobufConverter {
       final items = <pb.ServiceWithCharacteristics>[];
       for (final serviceId in servicesWithCharacteristicsToDiscover.keys) {
         final characteristicIds =
-            servicesWithCharacteristicsToDiscover[serviceId];
+            servicesWithCharacteristicsToDiscover[serviceId]!;
         items.add(
           pb.ServiceWithCharacteristics()
             ..serviceId = (pb.Uuid()..data = serviceId.data)
@@ -171,13 +170,13 @@ class ArgsToProtobufConverterImpl implements ArgsToProtobufConverter {
 
   @override
   pb.ScanForDevicesRequest createScanForDevicesRequest({
-    @required List<Uuid> withServices,
-    @required ScanMode scanMode,
-    @required bool requireLocationServicesEnabled,
+    required List<Uuid>? withServices,
+    required ScanMode? scanMode,
+    required bool? requireLocationServicesEnabled,
   }) {
     final args = pb.ScanForDevicesRequest()
-      ..scanMode = convertScanModeToArgs(scanMode)
-      ..requireLocationServicesEnabled = requireLocationServicesEnabled;
+      ..scanMode = convertScanModeToArgs(scanMode!)
+      ..requireLocationServicesEnabled = requireLocationServicesEnabled!;
 
     if (withServices != null) {
       for (final withService in withServices) {
