@@ -13,8 +13,8 @@ part 'discovered_device.g.dart';
 /// Result of a scan interval.
 @immutable
 class ScanResult {
-  final Result<DiscoveredDevice, GenericFailure<ScanFailure>> result;
-  const ScanResult({@required this.result});
+  final Result<DiscoveredDevice, GenericFailure<ScanFailure>?> result;
+  const ScanResult({required this.result});
 
   @override
   String toString() => "$ScanResult(result: $result)";
@@ -36,11 +36,11 @@ class DiscoveredDevice extends $DiscoveredDevice {
   final int rssi;
 
   const DiscoveredDevice({
-    @required this.id,
-    @required this.name,
-    @required this.serviceData,
-    @required this.manufacturerData,
-    @required this.rssi,
+    required this.id,
+    required this.name,
+    required this.serviceData,
+    required this.manufacturerData,
+    required this.rssi,
   });
 }
 
