@@ -1,7 +1,6 @@
 import 'package:collection/collection.dart';
 import 'package:flutter_reactive_ble/src/model/uuid.dart';
 import 'package:functional_data/functional_data.dart';
-import 'package:meta/meta.dart';
 
 part 'discovered_service.g.dart';
 //ignore_for_file: annotate_overrides
@@ -9,8 +8,8 @@ part 'discovered_service.g.dart';
 @FunctionalData()
 class DiscoveredService extends $DiscoveredService {
   const DiscoveredService({
-    @required this.serviceId,
-    @required this.characteristicIds,
+    required this.serviceId,
+    required this.characteristicIds,
     this.includedServices = const [],
   });
 
