@@ -1,10 +1,5 @@
-import 'package:meta/meta.dart';
-
 T selectFrom<T>(List<T> values,
-    {@required int index, @required T Function(int index) fallback}) {
-  assert(values != null);
-  assert(fallback != null);
-
+    {required int? index, required T Function(int? index) fallback}) {
   if (index != null && index >= 0 && index < values.length) {
     return values[index];
   }
