@@ -16,8 +16,8 @@ import 'connected_device_operation_test.mocks.dart';
 
 @GenerateMocks([DeviceOperationController])
 void main() {
-  MockDeviceOperationController _controller;
-  ConnectedDeviceOperation _sut;
+  late MockDeviceOperationController _controller;
+  late ConnectedDeviceOperation _sut;
 
   group('$ConnectedDeviceOperation', () {
     setUp(() {
@@ -27,7 +27,7 @@ void main() {
       );
     });
     group('Listen to char value updates', () {
-      CharacteristicValue valueUpdate;
+      CharacteristicValue? valueUpdate;
 
       setUp(() {
         valueUpdate = CharacteristicValue(
@@ -40,25 +40,25 @@ void main() {
         );
 
         when(_controller.charValueUpdateStream)
-            .thenAnswer((_) => Stream.fromIterable([valueUpdate]));
+            .thenAnswer((_) => Stream.fromIterable([valueUpdate!]));
       });
 
       test('It emits value updates received from plugincontroller', () {
         expect(
           _sut.characteristicValueStream,
-          emitsInOrder(<CharacteristicValue>[valueUpdate]),
+          emitsInOrder(<CharacteristicValue?>[valueUpdate]),
         );
       });
     });
 
     group('Read characteristic', () {
-      QualifiedCharacteristic charDevice;
+      late QualifiedCharacteristic charDevice;
       QualifiedCharacteristic charOtherSameDevice;
       QualifiedCharacteristic charOtherDevice;
-      CharacteristicValue valueUpdate;
-      CharacteristicValue valueUpdateOtherDevice;
-      CharacteristicValue valueUpdateSameDeviceOtherChar;
-      List<int> result;
+      CharacteristicValue? valueUpdate;
+      CharacteristicValue? valueUpdateOtherDevice;
+      CharacteristicValue? valueUpdateSameDeviceOtherChar;
+      List<int>? result;
 
       setUp(() {
         charDevice = QualifiedCharacteristic(
@@ -103,9 +103,9 @@ void main() {
         setUp(() async {
           when(_controller.charValueUpdateStream)
               .thenAnswer((_) => Stream.fromIterable([
-                    valueUpdate,
-                    valueUpdateOtherDevice,
-                    valueUpdateSameDeviceOtherChar,
+                    valueUpdate!,
+                    valueUpdateOtherDevice!,
+                    valueUpdateSameDeviceOtherChar!,
                   ]));
 
           result = await _sut.readCharacteristic(charDevice);
@@ -122,8 +122,8 @@ void main() {
         setUp(() async {
           when(_controller.charValueUpdateStream)
               .thenAnswer((_) => Stream.fromIterable([
-                    valueUpdateOtherDevice,
-                    valueUpdateSameDeviceOtherChar,
+                    valueUpdateOtherDevice!,
+                    valueUpdateSameDeviceOtherChar!,
                   ]));
         });
 
@@ -137,7 +137,7 @@ void main() {
     });
 
     group('Write characteristic', () {
-      QualifiedCharacteristic characteristic;
+      late QualifiedCharacteristic characteristic;
       WriteCharacteristicInfo info;
       const value = [1, 0];
 
@@ -155,7 +155,7 @@ void main() {
             info = WriteCharacteristicInfo(
               characteristic: characteristic,
               result: const Result<void,
-                  GenericFailure<WriteCharacteristicFailure>>.success(null),
+                  GenericFailure<WriteCharacteristicFailure>>.success(1),
             );
 
             when(_controller.writeCharacteristicWithResponse(any, any))
@@ -201,7 +201,7 @@ void main() {
               info = WriteCharacteristicInfo(
                 characteristic: characteristic,
                 result: const Result<void,
-                    GenericFailure<WriteCharacteristicFailure>>.success(null),
+                    GenericFailure<WriteCharacteristicFailure>>.success(1),
               );
 
               when(_controller.writeCharacteristicWithoutResponse(any, any))
@@ -245,15 +245,15 @@ void main() {
       });
 
       group('Subscribe to characteristic', () {
-        QualifiedCharacteristic charDevice;
+        late QualifiedCharacteristic charDevice;
         QualifiedCharacteristic charOtherSameDevice;
         QualifiedCharacteristic charOtherDevice;
-        CharacteristicValue valueUpdate1;
-        CharacteristicValue valueUpdate2;
-        CharacteristicValue valueUpdateOtherDevice;
-        CharacteristicValue valueUpdateSameDeviceOtherChar;
-        Stream<List<int>> result;
-        Completer<ConnectionStateUpdate> terminateCompleter;
+        late CharacteristicValue valueUpdate1;
+        late CharacteristicValue valueUpdate2;
+        late CharacteristicValue valueUpdateOtherDevice;
+        late CharacteristicValue valueUpdateSameDeviceOtherChar;
+        late Stream<List<int>?> result;
+        late Completer<ConnectionStateUpdate> terminateCompleter;
 
         setUp(() {
           terminateCompleter = Completer();
@@ -330,7 +330,7 @@ void main() {
       group('Negotiate mtusize', () {
         const deviceId = '123';
         const mtuSize = 50;
-        int result;
+        int? result;
 
         setUp(() async {
           when(_controller.requestMtuSize(any, any))
@@ -346,7 +346,7 @@ void main() {
 
       group('Change connection priority', () {
         const deviceId = '123';
-        ConnectionPriority priority;
+        late ConnectionPriority priority;
 
         setUp(() {
           priority = ConnectionPriority.highPerformance;
@@ -356,7 +356,7 @@ void main() {
           setUp(() {
             when(_controller.requestConnectionPriority(any, any))
                 .thenAnswer((_) async => const ConnectionPriorityInfo(
-                      result: Result.success(null),
+                      result: Result.success(1),
                     ));
           });
 
