@@ -2,28 +2,16 @@
 //  Generated code. Do not modify.
 //  source: bledata.proto
 //
-// @dart = 2.3
+// @dart = 2.12
 // ignore_for_file: annotate_overrides,camel_case_types,unnecessary_const,non_constant_identifier_names,library_prefixes,unused_import,unused_shown_name,return_of_invalid_type,unnecessary_this,prefer_final_fields
 
+import 'dart:core' as $core;
 const ScanForDevicesRequest$json = const {
   '1': 'ScanForDevicesRequest',
   '2': const [
-    const {
-      '1': 'serviceUuids',
-      '3': 1,
-      '4': 3,
-      '5': 11,
-      '6': '.Uuid',
-      '10': 'serviceUuids'
-    },
+    const {'1': 'serviceUuids', '3': 1, '4': 3, '5': 11, '6': '.Uuid', '10': 'serviceUuids'},
     const {'1': 'scanMode', '3': 2, '4': 1, '5': 5, '10': 'scanMode'},
-    const {
-      '1': 'requireLocationServicesEnabled',
-      '3': 3,
-      '4': 1,
-      '5': 8,
-      '10': 'requireLocationServicesEnabled'
-    },
+    const {'1': 'requireLocationServicesEnabled', '3': 3, '4': 1, '5': 8, '10': 'requireLocationServicesEnabled'},
   ],
 };
 
@@ -32,29 +20,9 @@ const DeviceScanInfo$json = const {
   '2': const [
     const {'1': 'id', '3': 1, '4': 1, '5': 9, '10': 'id'},
     const {'1': 'name', '3': 2, '4': 1, '5': 9, '10': 'name'},
-    const {
-      '1': 'failure',
-      '3': 3,
-      '4': 1,
-      '5': 11,
-      '6': '.GenericFailure',
-      '10': 'failure'
-    },
-    const {
-      '1': 'serviceData',
-      '3': 4,
-      '4': 3,
-      '5': 11,
-      '6': '.ServiceDataEntry',
-      '10': 'serviceData'
-    },
-    const {
-      '1': 'manufacturerData',
-      '3': 6,
-      '4': 1,
-      '5': 12,
-      '10': 'manufacturerData'
-    },
+    const {'1': 'failure', '3': 3, '4': 1, '5': 11, '6': '.GenericFailure', '10': 'failure'},
+    const {'1': 'serviceData', '3': 4, '4': 3, '5': 11, '6': '.ServiceDataEntry', '10': 'serviceData'},
+    const {'1': 'manufacturerData', '3': 6, '4': 1, '5': 12, '10': 'manufacturerData'},
     const {'1': 'rssi', '3': 5, '4': 1, '5': 5, '10': 'rssi'},
   ],
 };
@@ -63,14 +31,7 @@ const ConnectToDeviceRequest$json = const {
   '1': 'ConnectToDeviceRequest',
   '2': const [
     const {'1': 'deviceId', '3': 1, '4': 1, '5': 9, '10': 'deviceId'},
-    const {
-      '1': 'servicesWithCharacteristicsToDiscover',
-      '3': 2,
-      '4': 1,
-      '5': 11,
-      '6': '.ServicesWithCharacteristics',
-      '10': 'servicesWithCharacteristicsToDiscover'
-    },
+    const {'1': 'servicesWithCharacteristicsToDiscover', '3': 2, '4': 1, '5': 11, '6': '.ServicesWithCharacteristics', '10': 'servicesWithCharacteristicsToDiscover'},
     const {'1': 'timeoutInMs', '3': 3, '4': 1, '5': 5, '10': 'timeoutInMs'},
   ],
 };
@@ -79,21 +40,8 @@ const DeviceInfo$json = const {
   '1': 'DeviceInfo',
   '2': const [
     const {'1': 'id', '3': 1, '4': 1, '5': 9, '10': 'id'},
-    const {
-      '1': 'connectionState',
-      '3': 2,
-      '4': 1,
-      '5': 5,
-      '10': 'connectionState'
-    },
-    const {
-      '1': 'failure',
-      '3': 3,
-      '4': 1,
-      '5': 11,
-      '6': '.GenericFailure',
-      '10': 'failure'
-    },
+    const {'1': 'connectionState', '3': 2, '4': 1, '5': 5, '10': 'connectionState'},
+    const {'1': 'failure', '3': 3, '4': 1, '5': 11, '6': '.GenericFailure', '10': 'failure'},
   ],
 };
 
@@ -114,93 +62,44 @@ const ClearGattCacheRequest$json = const {
 const ClearGattCacheInfo$json = const {
   '1': 'ClearGattCacheInfo',
   '2': const [
-    const {
-      '1': 'failure',
-      '3': 1,
-      '4': 1,
-      '5': 11,
-      '6': '.GenericFailure',
-      '10': 'failure'
-    },
+    const {'1': 'failure', '3': 1, '4': 1, '5': 11, '6': '.GenericFailure', '10': 'failure'},
   ],
 };
 
 const NotifyCharacteristicRequest$json = const {
   '1': 'NotifyCharacteristicRequest',
   '2': const [
-    const {
-      '1': 'characteristic',
-      '3': 1,
-      '4': 1,
-      '5': 11,
-      '6': '.CharacteristicAddress',
-      '10': 'characteristic'
-    },
+    const {'1': 'characteristic', '3': 1, '4': 1, '5': 11, '6': '.CharacteristicAddress', '10': 'characteristic'},
   ],
 };
 
 const NotifyNoMoreCharacteristicRequest$json = const {
   '1': 'NotifyNoMoreCharacteristicRequest',
   '2': const [
-    const {
-      '1': 'characteristic',
-      '3': 1,
-      '4': 1,
-      '5': 11,
-      '6': '.CharacteristicAddress',
-      '10': 'characteristic'
-    },
+    const {'1': 'characteristic', '3': 1, '4': 1, '5': 11, '6': '.CharacteristicAddress', '10': 'characteristic'},
   ],
 };
 
 const ReadCharacteristicRequest$json = const {
   '1': 'ReadCharacteristicRequest',
   '2': const [
-    const {
-      '1': 'characteristic',
-      '3': 1,
-      '4': 1,
-      '5': 11,
-      '6': '.CharacteristicAddress',
-      '10': 'characteristic'
-    },
+    const {'1': 'characteristic', '3': 1, '4': 1, '5': 11, '6': '.CharacteristicAddress', '10': 'characteristic'},
   ],
 };
 
 const CharacteristicValueInfo$json = const {
   '1': 'CharacteristicValueInfo',
   '2': const [
-    const {
-      '1': 'characteristic',
-      '3': 1,
-      '4': 1,
-      '5': 11,
-      '6': '.CharacteristicAddress',
-      '10': 'characteristic'
-    },
+    const {'1': 'characteristic', '3': 1, '4': 1, '5': 11, '6': '.CharacteristicAddress', '10': 'characteristic'},
     const {'1': 'value', '3': 2, '4': 1, '5': 12, '10': 'value'},
-    const {
-      '1': 'failure',
-      '3': 3,
-      '4': 1,
-      '5': 11,
-      '6': '.GenericFailure',
-      '10': 'failure'
-    },
+    const {'1': 'failure', '3': 3, '4': 1, '5': 11, '6': '.GenericFailure', '10': 'failure'},
   ],
 };
 
 const WriteCharacteristicRequest$json = const {
   '1': 'WriteCharacteristicRequest',
   '2': const [
-    const {
-      '1': 'characteristic',
-      '3': 1,
-      '4': 1,
-      '5': 11,
-      '6': '.CharacteristicAddress',
-      '10': 'characteristic'
-    },
+    const {'1': 'characteristic', '3': 1, '4': 1, '5': 11, '6': '.CharacteristicAddress', '10': 'characteristic'},
     const {'1': 'value', '3': 2, '4': 1, '5': 12, '10': 'value'},
   ],
 };
@@ -208,22 +107,8 @@ const WriteCharacteristicRequest$json = const {
 const WriteCharacteristicInfo$json = const {
   '1': 'WriteCharacteristicInfo',
   '2': const [
-    const {
-      '1': 'characteristic',
-      '3': 1,
-      '4': 1,
-      '5': 11,
-      '6': '.CharacteristicAddress',
-      '10': 'characteristic'
-    },
-    const {
-      '1': 'failure',
-      '3': 3,
-      '4': 1,
-      '5': 11,
-      '6': '.GenericFailure',
-      '10': 'failure'
-    },
+    const {'1': 'characteristic', '3': 1, '4': 1, '5': 11, '6': '.CharacteristicAddress', '10': 'characteristic'},
+    const {'1': 'failure', '3': 3, '4': 1, '5': 11, '6': '.GenericFailure', '10': 'failure'},
   ],
 };
 
@@ -240,14 +125,7 @@ const NegotiateMtuInfo$json = const {
   '2': const [
     const {'1': 'deviceId', '3': 1, '4': 1, '5': 9, '10': 'deviceId'},
     const {'1': 'mtuSize', '3': 2, '4': 1, '5': 5, '10': 'mtuSize'},
-    const {
-      '1': 'failure',
-      '3': 3,
-      '4': 1,
-      '5': 11,
-      '6': '.GenericFailure',
-      '10': 'failure'
-    },
+    const {'1': 'failure', '3': 3, '4': 1, '5': 11, '6': '.GenericFailure', '10': 'failure'},
   ],
 };
 
@@ -270,14 +148,7 @@ const ChangeConnectionPriorityInfo$json = const {
   '1': 'ChangeConnectionPriorityInfo',
   '2': const [
     const {'1': 'deviceId', '3': 1, '4': 1, '5': 9, '10': 'deviceId'},
-    const {
-      '1': 'failure',
-      '3': 2,
-      '4': 1,
-      '5': 11,
-      '6': '.GenericFailure',
-      '10': 'failure'
-    },
+    const {'1': 'failure', '3': 2, '4': 1, '5': 11, '6': '.GenericFailure', '10': 'failure'},
   ],
 };
 
@@ -285,36 +156,15 @@ const CharacteristicAddress$json = const {
   '1': 'CharacteristicAddress',
   '2': const [
     const {'1': 'deviceId', '3': 1, '4': 1, '5': 9, '10': 'deviceId'},
-    const {
-      '1': 'serviceUuid',
-      '3': 2,
-      '4': 1,
-      '5': 11,
-      '6': '.Uuid',
-      '10': 'serviceUuid'
-    },
-    const {
-      '1': 'characteristicUuid',
-      '3': 3,
-      '4': 1,
-      '5': 11,
-      '6': '.Uuid',
-      '10': 'characteristicUuid'
-    },
+    const {'1': 'serviceUuid', '3': 2, '4': 1, '5': 11, '6': '.Uuid', '10': 'serviceUuid'},
+    const {'1': 'characteristicUuid', '3': 3, '4': 1, '5': 11, '6': '.Uuid', '10': 'characteristicUuid'},
   ],
 };
 
 const ServiceDataEntry$json = const {
   '1': 'ServiceDataEntry',
   '2': const [
-    const {
-      '1': 'serviceUuid',
-      '3': 1,
-      '4': 1,
-      '5': 11,
-      '6': '.Uuid',
-      '10': 'serviceUuid'
-    },
+    const {'1': 'serviceUuid', '3': 1, '4': 1, '5': 11, '6': '.Uuid', '10': 'serviceUuid'},
     const {'1': 'data', '3': 2, '4': 1, '5': 12, '10': 'data'},
   ],
 };
@@ -322,36 +172,15 @@ const ServiceDataEntry$json = const {
 const ServicesWithCharacteristics$json = const {
   '1': 'ServicesWithCharacteristics',
   '2': const [
-    const {
-      '1': 'items',
-      '3': 1,
-      '4': 3,
-      '5': 11,
-      '6': '.ServiceWithCharacteristics',
-      '10': 'items'
-    },
+    const {'1': 'items', '3': 1, '4': 3, '5': 11, '6': '.ServiceWithCharacteristics', '10': 'items'},
   ],
 };
 
 const ServiceWithCharacteristics$json = const {
   '1': 'ServiceWithCharacteristics',
   '2': const [
-    const {
-      '1': 'serviceId',
-      '3': 1,
-      '4': 1,
-      '5': 11,
-      '6': '.Uuid',
-      '10': 'serviceId'
-    },
-    const {
-      '1': 'characteristics',
-      '3': 2,
-      '4': 3,
-      '5': 11,
-      '6': '.Uuid',
-      '10': 'characteristics'
-    },
+    const {'1': 'serviceId', '3': 1, '4': 1, '5': 11, '6': '.Uuid', '10': 'serviceId'},
+    const {'1': 'characteristics', '3': 2, '4': 3, '5': 11, '6': '.Uuid', '10': 'characteristics'},
   ],
 };
 
@@ -366,44 +195,16 @@ const DiscoverServicesInfo$json = const {
   '1': 'DiscoverServicesInfo',
   '2': const [
     const {'1': 'deviceId', '3': 1, '4': 1, '5': 9, '10': 'deviceId'},
-    const {
-      '1': 'services',
-      '3': 2,
-      '4': 3,
-      '5': 11,
-      '6': '.DiscoveredService',
-      '10': 'services'
-    },
+    const {'1': 'services', '3': 2, '4': 3, '5': 11, '6': '.DiscoveredService', '10': 'services'},
   ],
 };
 
 const DiscoveredService$json = const {
   '1': 'DiscoveredService',
   '2': const [
-    const {
-      '1': 'serviceUuid',
-      '3': 1,
-      '4': 1,
-      '5': 11,
-      '6': '.Uuid',
-      '10': 'serviceUuid'
-    },
-    const {
-      '1': 'characteristicUuids',
-      '3': 2,
-      '4': 3,
-      '5': 11,
-      '6': '.Uuid',
-      '10': 'characteristicUuids'
-    },
-    const {
-      '1': 'includedServices',
-      '3': 3,
-      '4': 3,
-      '5': 11,
-      '6': '.DiscoveredService',
-      '10': 'includedServices'
-    },
+    const {'1': 'serviceUuid', '3': 1, '4': 1, '5': 11, '6': '.Uuid', '10': 'serviceUuid'},
+    const {'1': 'characteristicUuids', '3': 2, '4': 3, '5': 11, '6': '.Uuid', '10': 'characteristicUuids'},
+    const {'1': 'includedServices', '3': 3, '4': 3, '5': 11, '6': '.DiscoveredService', '10': 'includedServices'},
   ],
 };
 
@@ -421,3 +222,4 @@ const GenericFailure$json = const {
     const {'1': 'message', '3': 2, '4': 1, '5': 9, '10': 'message'},
   ],
 };
+
