@@ -14,11 +14,11 @@ abstract class $ConnectionStateUpdate {
   const $ConnectionStateUpdate();
   String get deviceId;
   DeviceConnectionState get connectionState;
-  GenericFailure<ConnectionError> get failure;
+  GenericFailure<ConnectionError>? get failure;
   ConnectionStateUpdate copyWith(
-          {String deviceId,
-          DeviceConnectionState connectionState,
-          GenericFailure<ConnectionError> failure}) =>
+          {String? deviceId,
+          DeviceConnectionState? connectionState,
+          GenericFailure<ConnectionError>? failure}) =>
       ConnectionStateUpdate(
           deviceId: deviceId ?? this.deviceId,
           connectionState: connectionState ?? this.connectionState,
@@ -27,7 +27,8 @@ abstract class $ConnectionStateUpdate {
   String toString() =>
       "ConnectionStateUpdate(deviceId: $deviceId, connectionState: $connectionState, failure: $failure)";
   @override
-  bool operator ==(dynamic other) =>
+  bool operator ==(Object other) =>
+      other is ConnectionStateUpdate &&
       other.runtimeType == runtimeType &&
       deviceId == other.deviceId &&
       connectionState == other.connectionState &&
@@ -51,6 +52,6 @@ class ConnectionStateUpdate$ {
           (s_, connectionState) =>
               s_.copyWith(connectionState: connectionState));
   static final failure =
-      Lens<ConnectionStateUpdate, GenericFailure<ConnectionError>>(
+      Lens<ConnectionStateUpdate, GenericFailure<ConnectionError>?>(
           (s_) => s_.failure, (s_, failure) => s_.copyWith(failure: failure));
 }
