@@ -6,7 +6,7 @@ void main() {
     test("executes the failure branch for a non-null failure object", () {
       const failureObject = 1;
       const sut = Result<int, int>.failure(failureObject);
-      final handler = _ResultHandler<int, int, int>();
+      final handler = _ResultHandler<int?, int?, int>();
 
       final result =
           sut.iif(success: handler.success, failure: handler.failure);
@@ -17,7 +17,7 @@ void main() {
     test("executes the success branch for a non-null success value", () {
       const value = 1;
       const sut = Result<int, int>.success(value);
-      final handler = _ResultHandler<int, int, int>();
+      final handler = _ResultHandler<int?, int?, int>();
 
       final result =
           sut.iif(success: handler.success, failure: handler.failure);
@@ -26,9 +26,9 @@ void main() {
     });
 
     test("executes the success branch for null success value", () {
-      const int value = null;
-      const sut = Result<int, int>.success(value);
-      final handler = _ResultHandler<int, int, int>();
+      const int? value = null;
+      const sut = Result<int, int?>.success(value);
+      final handler = _ResultHandler<int?, int?, int?>();
 
       final result =
           sut.iif(success: handler.success, failure: handler.failure);
